pyproject.toml
[project]
name = "invoice-review-openenv"
version = "1.0.0"
description = "Invoice Review OpenEnv benchmark environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "openai>=1.6.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
